leetcode Find Polygon With the Largest Perimeter - 2차 풀이 완료 - 자료형과 값을 확인할 것

+++ CodingTest/LeetCode/Quiz.java
@@ -29,11 +29,11 @@ class Solution {
     public long largestPerimeter(final int[] nums) {
         Arrays.sort(nums);
 
-        int ans = -1;
+        long ans = -1;
 
         //처음에 3개로 값을 구해보고, 계속 조금씩 더하면서 폴리곤이 만들어지는지 체크하고
         //폴리곤이 만들어지면 둘레 값을 구한다.
-        int preSum = nums[0] + nums[1];
+        long preSum = (long) nums[0] + nums[1];
         for (int i = 2; i < nums.length; i++) {
             final int val = nums[i];
